+++ libtls-sys/src/lib.rs
@@ -22,8 +22,10 @@
 
 include!(concat!(env!("OUT_DIR"), "/bindings.rs"));
 
-use std::marker::{Send, Sync};
-use std::os::raw::c_int;
+use std::{
+    marker::{Send, Sync},
+    os::raw::c_int,
+};
 
 extern "C" {
     pub fn close(fd: c_int) -> c_int;